README: fix stale .NET Standard link
The "master" branch on https://github.com/dotnet/standard does not exist anymore.
Github redirects that link to the current default branch "archive". But since
the expected file does not exist on that branch, users are faced with a 404 page.
Link to the "release/3.0" branch instead, which seems to have the latest version
of the document.