A simple TreeView sample project
I was able to get the TreeView to populate from a folder with all it's subfolders. This is from a VS 2022 project and not a JetBrains one if that is important. The code also may not be up to par with what you all normally publish but this is a pretty non-complicated example if you are interested.